src/student/requirements.txt
pystray==0.19.5
Pillow==10.4.0
mss==9.0.2
pywin32==308
requests==2.32.3
google-auth-oauthlib==1.2.1
google-auth==2.38.0
pyinstaller==6.11.1
win10toast==0.9